bsr(orga) move static list to component
