+++ orga/app/components/sco-organization-participant/list.hbs
@@ -27,7 +27,7 @@
 
   <Ui::SelectFilter
     @field="connexionType"
-    @options={{@connectionTypesOptions}}
+    @options={{this.connectionTypesOptions}}
     @selectedOption={{@connexionTypeFilter}}
     @triggerFiltering={{@onFilter}}
     @ariaLabel={{t "pages.sco-organization-participants.filter.login-method.aria-label"}}
@@ -38,7 +38,7 @@
     @title={{t "pages.sco-organization-participants.filter.certificability.label"}}
     @onSelect={{@onFilter}}
     @selectedOption={{@certificabilityFilter}}
-    @options={{@certificabilityOptions}}
+    @options={{this.certificabilityOptions}}
     @placeholder={{t "pages.sco-organization-participants.filter.certificability.label"}}
     @ariaLabel={{t "pages.sco-organization-participants.filter.certificability.aria-label"}}
     @emptyMessage={{t "pages.sco-organization-participants.filter.certificability.empty"}}
