+++ orga/app/components/sco-organization-participant/list.js
@@ -2,12 +2,38 @@ import { action } from '@ember/object';
 import { inject as service } from '@ember/service';
 import Component from '@glimmer/component';
 import { tracked } from '@glimmer/tracking';
-
+import { CONNECTION_TYPES } from '../../models/sco-organization-participant';
 export default class ScoList extends Component {
   @service currentUser;
-
+  @service intl;
   @tracked student = null;
   @tracked isShowingAuthenticationMethodModal = false;
+
+  get connectionTypesOptions() {
+    return [
+      { value: 'none', label: this.intl.t(CONNECTION_TYPES.none) },
+      { value: 'email', label: this.intl.t(CONNECTION_TYPES.email) },
+      { value: 'identifiant', label: this.intl.t(CONNECTION_TYPES.identifiant) },
+      { value: 'mediacentre', label: this.intl.t(CONNECTION_TYPES.mediacentre) },
+    ];
+  }
+
+  get certificabilityOptions() {
+    return [
+      {
+        value: 'not-available',
+        label: this.intl.t('pages.sco-organization-participants.table.column.is-certifiable.not-available'),
+      },
+      {
+        value: 'eligible',
+        label: this.intl.t('pages.sco-organization-participants.table.column.is-certifiable.eligible'),
+      },
+      {
+        value: 'non-eligible',
+        label: this.intl.t('pages.sco-organization-participants.table.column.is-certifiable.non-eligible'),
+      },
+    ];
+  }
 
   loadDivisions = async () => {
     const divisions = await this.currentUser.organization.divisions;
