+++ orga/app/controllers/authenticated/sco-organization-participants/list.js
@@ -2,8 +2,6 @@ import { action } from '@ember/object';
 import { inject as service } from '@ember/service';
 import Controller from '@ember/controller';
 import { tracked } from '@glimmer/tracking';
-import { CONNECTION_TYPES } from '../../../models/sco-organization-participant';
-
 export default class ListController extends Controller {
   @service currentUser;
   @service notifications;
@@ -33,32 +31,6 @@ export default class ListController extends Controller {
     this.connexionType = null;
     this.search = null;
     this.certificability = [];
-  }
-
-  get connectionTypesOptions() {
-    return [
-      { value: 'none', label: this.intl.t(CONNECTION_TYPES.none) },
-      { value: 'email', label: this.intl.t(CONNECTION_TYPES.email) },
-      { value: 'identifiant', label: this.intl.t(CONNECTION_TYPES.identifiant) },
-      { value: 'mediacentre', label: this.intl.t(CONNECTION_TYPES.mediacentre) },
-    ];
-  }
-
-  get certificabilityOptions() {
-    return [
-      {
-        value: 'not-available',
-        label: this.intl.t('pages.sco-organization-participants.table.column.is-certifiable.not-available'),
-      },
-      {
-        value: 'eligible',
-        label: this.intl.t('pages.sco-organization-participants.table.column.is-certifiable.eligible'),
-      },
-      {
-        value: 'non-eligible',
-        label: this.intl.t('pages.sco-organization-participants.table.column.is-certifiable.non-eligible'),
-      },
-    ];
   }
 
   @action
