+++ orga/app/templates/authenticated/sco-organization-participants/list.hbs
@@ -7,8 +7,6 @@
   />
   <ScoOrganizationParticipant::List
     @students={{@model}}
-    @connectionTypesOptions={{this.connectionTypesOptions}}
-    @certificabilityOptions={{this.certificabilityOptions}}
     @searchFilter={{this.search}}
     @divisionsFilter={{this.divisions}}
     @connexionTypeFilter={{this.connexionType}}
